ocr_project/requirements.txt
easyocr>=1.7.0
opencv-python>=4.5.0
Pillow>=9.0.0
